Only offer gtk options for gnucash, not for gnucash-cli
* As the gtk command line options are only relevant for
  gnucash and not gnucash-cli, remove the bit that adds
  them from Gnucash::CoreApp.
* As the gtk command line options are parsed out by gtk_init_check
  it's sufficient to run that before parsing our own command line
  options to have gtk handle them for us.
* That just leaves us with the loss of a help message for the
  gtk command line options. To handle that we derive a class
  Gnucash::Gnucash from Gnucash::CoreApp which will generate an
  option context for the gtk options purely to extract the help message.
  This will then be used if the users pass option --help-gtk.

This is a bit clumsy to do with GOptionContext as it requires to store
a few gnucash only parameters in Gnucash::CoreApp. The plan is to
improve this in a future commit.
Note however we will still be stuck with the dummy option context generation
for the gtk option help message. There's just no other way to extract this
message otherwise.